+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [0.3.1] - 2022-01-10
+
 - Add `CallableTree::Node::Internal#seek?` to check whether the node's strategy is `seek` or not.
 - Add `CallableTree::Node::Internal#broadcast?` to check whether the node's strategy is `broadcast` or not.
 - Add `CallableTree::Node::Internal#compose?` to check whether the node's strategy is `compose` or not.